+++ src/main/java/br/com/aviapp/api/application/dto/DailyReportDTO.java
@@ -1,14 +1,13 @@
 package br.com.aviapp.api.application.dto;
 
-import br.com.aviapp.api.domain.entities.AviaryReportBO;
-
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
 public record DailyReportDTO(Date date, List<AviaryReportDTO> aviaryReports, int totalEggsCollected, int totalDeadBirds,
                              int totalDeadChickens, int totalDeadRoosters, int currentChickens, int currentRoosters,
-                             int totalBirds, float production,
-                             float roosterMortality, float chickenMortality, float mortality,
-                             float chickenRoosterProportion, List<EggDetailDTO> quantityByEggType) {
+                             int totalBirds, BigDecimal production,
+                             BigDecimal roosterMortality, BigDecimal chickenMortality, BigDecimal mortality,
+                             BigDecimal chickenRoosterProportion, List<EggDetailDTO> quantityByEggType) {
 }
