+++ src/main/java/br/com/aviapp/api/domain/entities/DailyReportBO.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -23,11 +24,11 @@ public class DailyReportBO {
     private int currentChickens;
     private int currentRoosters;
     private int totalBirds;
-    private float production;
-    private float roosterMortality;
-    private float chickenMortality;
-    private float mortality;
-    private float chickenRoosterProportion;
+    private BigDecimal production;
+    private BigDecimal roosterMortality;
+    private BigDecimal chickenMortality;
+    private BigDecimal mortality;
+    private BigDecimal chickenRoosterProportion;
     private List<EggDetailBO> quantityByEggType;
